cohesive themes (pureline, tmux, vim)
I have tweaked the colour configs in the tmux and pureline configs to
match the color theme for my vimrc. Everything is consistent now, and it
looks beautiful.

I have also done away with vim-plug, and instead just leverage native
pack functionality for plugins. Each plugin is a submodule in the .vim
directory, and I have updated the Makefile accordingly.