no more AutoComplPop
If plugin (written 10+ years ago) can be accomplished in 18-or-so lines
of vimscript in your vimrc, why use it?

Wrote a cute autocmd function executed on InsertCharPre that feeds the
autocomplete popup keys automatically. I've found that this actually
works better than the plugin, and it's actually pretty efficient (even
with tag completion). I have also renapped <CR> to exit the
autocomplete, which is nice when I just want to insert a damn CR.

I also changed the terminal keymaps to <C-s>v and <C-s>h, for vertical
and horizontal splits. Makes a bit more sense.