a few fixes n stuff
Theres a particularly annoying bug in Vim where completion will enter an
infinite loop when I type too quickly. There's an issue logged but it
was closed. A workaround is to execute findkeys() in a timer_start(). It
has been working nicely now, so I'm keeping it.

A few other changes as well, here are there.