add mapping for cleaning up buffers
I tend to accumulate a lot of buffers when working in Vim, and they
clutter my tabline and annoy me. I introduced a mapping that removes any
buffers that aren't displayed in any windows. The buffers still hang
around, but are unlisted. This is pretty useful to ensure my completion
still behaves as it should.