completion fixes, fern changes
When I'm typing some text in the middle of a line of code, the
completion menu wasn't showing. I found that I didn't slice the buffer
line properly (off by one), so I was getting the next character (which
is often not a keyword character). I corrected this and it works again,
and it's amazinggggggg.

I also added the -keep option in my <C-e> mapping when opening the fern
drawer, because I often find myself accidentally closing vim when I
close the last buffer.