create mapping to toggle quickfix window
The default behaviour for the quickfix window is pretty ok-ish, but can
be improved. I added a mapping to open the qf window in the same way as
the terminal (dock). The mapping will also toggle the window.

I also added mappings for going to the next quickfix item (cnext/cprev).