tag navigation improvements
When there are multiple occurrences of a keyword, <C-]> just takes you
to the first one. It's pretty rare that there's only one tag occurrence,
especially in larger projects. `:tjump` solves this, but the UX isn't
great.

Overrode the <C-]> mapping with a custom function that shows a simple
popup whenever there are multiple tag results.

Updated the fern plugin.