powerful tag-less navigation
We are gettin wild in here. If there's no tagfile present, replace the
tagfunc with one that returns returns from the current buffer. Handy
when you're trying to jump around in a file but don't have a tagfile.

I'll play around with this for a while and probably make adjustments.