fix minor bug in ins-completion
There was a weird bug in the auto ins-completion script that prevented
results from showing up correctly in certain cases. It was caused by
naive usage of pumvisible() to determine if completion is ongoing. If
there's no results, pumvisible() returns false but complete_info() can
be used to check if the completion is still ongoing.