taking the tabline intergalactic
I love my tabline setup, but it had one annoying flaw. The tabline fills
up quickly when you have a lot of buffers open, and when there are too
many buffers, the beginning of the tabline is truncated. It looks pretty
ugly when that happens, and sometimes you can't see the curent buffer
because it's in that part that is truncated.

This new revision takes the tabline to a whole new level. When there are
too many buffers and we can't show them all in the tabline, the extras
are dropped. If they can't be shown, a pretty scroll marker is shown.
Additionally, we try to keep the current buffer centered so that it's
always visible.

She's a beaut, Clark.