overhaul
A bunch of cleanup. Renaming functions for consistency, moving stuff to
autoload directories. I also made a few improvements to the tag
navigation features. Now the popup looks a bit more pretty.

I also added the vim tag file to the 'tags' option for the vim filetype.
This makes it easier to complete text and navigate to the help page for
something while working on vim files.