chore: switch to nord theme
Switched back to the nord theme after a year or so of using onedark.
Needed a change. This turned out to be a good switch, because it allowed
me to improve the consistency in the color palette between vim, bash and
tmux. All colours across all configurations are using the nord palette
now. And it looks beautiful.

I also cleaned up the theme configuration in my vim config. This is
something I've wanted to do for a while. Before, I just haphazardly
created highlights that somewhat matched the onedark theme. Now, any
highlights I use in my config derive from specific nord palette colors.