chore: hello vim-signature!
Let's give this a try. I've wanted to get better at using marks in Vim,
and displaying them in the sign column is a good way to get started. The
plugin is quite old and no longer maintained it seems, but it's
lightweight and works well.

A few minor interface changes too.