chore: bash prompt overhaul
Well, pureline, it's been a slice. You've been great, but it's time for
something greater.

The new bash prompt is based off of an example I found in the nord theme
ecosystem. Shamelessly stole that implementation and tweaked it to my
liking. Very clean :-)

There will probably be a lot of changes coming as I tweak and clean up
the implementation. For instance, the inputrc config will need to be
cleaned up somehow. We'll cross that bridge at another date.