chore: minor tabline tweaks
The fancy powerline characters don't look good on all systems. The look
is really dependent on the font size, and the wrong size will make the
tabline look pretty terrible.

Box drawing characters to the rescue! They always look good. It also
makes it look a little more consistent with the rest of the interface.