chore: next batch up of updates
Minor tweaks here. The most notable is a mapping for showing git blame
data in the quickfix list. Unfortunately because Git doesn't detect if
the terminal is a TTY, colored output mangles the output so it doesn't
work properly unless coloring is not configured. I'll fix this at some
point.