chore: checktime on certain autocmd events to trigger autoread
The 'autoread' setting will only reload files when changes from the
outside in very specific circumstances. In practice, this is almost
never triggered and I often find myself reloading the buffer with ':e'
(typically when a source formatter runs at the command line).

Add a simple autocmd for CursorHold, CursorHoldI and WinEnter to trigger
the file reload.